# Shorten and optimize let/endlet functions