--- a/examples/compiled/all-chars.sh
+++ b/examples/compiled/all-chars.sh
@@ -18,15 +18,15 @@ _main() {
 __=0
 __SP=0
 let() { # $1: variable name, $2: value (optional)
-  : $((__SP += 1)) $((__$__SP=$1)) # Push
-  : $(($1=${2-0}))                 # Init
+  : $((__$((__SP += 1))=$1)) # Push
+  : $(($1=${2-0}))           # Init
 }
 endlet() { # $1: return variable
            # $2...: function local variables
   __ret=$1 # Don't overwrite return value
   : $((__tmp = $__ret))
   while [ $# -ge 2 ]; do
-    : $(($2 = __$__SP)) $((__SP -= 1)); # Pop
+    : $(($2 = __$(((__SP -= 1) + 1)))) # Pop
     shift;
   done
   : $(($__ret=__tmp))   # Restore return value