## Context

The `let` and `endlet` functions are one of the most called functions in scripts generated by pnut. I noticed that the increment operation could be inlined in the push operation, which should be faster as each reference to a shell variable is an environment lookup which can be expensive on certain shells. The same optimization can be done for the pop operation in `endlet`.

## Benchmarks

Because zsh is so slow, I benchmarked `pnut-sh.sh` compiling `examples/sha256sum.c`.

```
Benchmark 1: dash ./build/pnut-sh.sh examples/sha256sum.c > out
  Time (mean ± σ):      1.913 s ±  0.032 s    [User: 1.707 s, System: 0.199 s]
  Range (min … max):    1.875 s …  1.963 s    10 runs
 
Benchmark 2: dash ./build/pnut-sh-fast.sh examples/sha256sum.c > out
  Time (mean ± σ):      1.872 s ±  0.021 s    [User: 1.671 s, System: 0.197 s]
  Range (min … max):    1.841 s …  1.908 s    10 runs
 
Result: 1.02 ± 0.02 times faster

Benchmark 1: ksh ./build/pnut-sh.sh examples/sha256sum.c > out
  Time (mean ± σ):      1.259 s ±  0.010 s    [User: 1.237 s, System: 0.019 s]
  Range (min … max):    1.251 s …  1.282 s    10 runs
 
Benchmark 2: ksh ./build/pnut-sh-fast.sh examples/sha256sum.c > out
  Time (mean ± σ):      1.251 s ±  0.014 s    [User: 1.225 s, System: 0.020 s]
  Range (min … max):    1.239 s …  1.280 s    10 runs
 
Result: 1.01 ± 0.01 times faster

Benchmark 1: yash ./build/pnut-sh.sh examples/sha256sum.c > out
  Time (mean ± σ):      8.229 s ±  0.099 s    [User: 4.993 s, System: 3.219 s]
  Range (min … max):    8.084 s …  8.377 s    10 runs
 
Benchmark 2: yash ./build/pnut-sh-fast.sh examples/sha256sum.c > out
  Time (mean ± σ):      8.064 s ±  0.090 s    [User: 4.820 s, System: 3.226 s]
  Range (min … max):    7.912 s …  8.216 s    10 runs
 
Result: 1.02 ± 0.02 times faster

Benchmark 1: bash ./build/pnut-sh.sh examples/sha256sum.c > out
  Time (mean ± σ):      4.235 s ±  0.042 s    [User: 4.186 s, System: 0.044 s]
  Range (min … max):    4.188 s …  4.291 s    10 runs
 
Benchmark 2: bash ./build/pnut-sh-fast.sh examples/sha256sum.c > out
  Time (mean ± σ):      4.252 s ±  0.072 s    [User: 4.192 s, System: 0.047 s]
  Range (min … max):    4.172 s …  4.408 s    10 runs
 
Result: 1.00 ± 0.02 times faster

Benchmark 1: zsh ./build/pnut-sh.sh examples/sha256sum.c > out
  Time (mean ± σ):     13.102 s ±  0.133 s    [User: 12.551 s, System: 0.505 s]
  Range (min … max):   12.925 s … 13.337 s    10 runs
 
Benchmark 2: zsh ./build/pnut-sh-fast.sh examples/sha256sum.c > out
  Time (mean ± σ):     13.114 s ±  0.155 s    [User: 12.566 s, System: 0.499 s]
  Range (min … max):   12.923 s … 13.345 s    10 runs
 
Result: 1.00 ± 0.02 times faster
```